Some db migrations, changed roles, changed some entities,added new methods to IdentityService, changed login model, now administrator will be register users, but no else, can't delete admin if he's alone, but he can delete himselft, aslo don't know what to do with db relative data about users, will know soon...

+++ StudentsEducation/Areas/Admin/Pages/Roles/Delete.cshtml.cs
@@ -40,7 +40,8 @@ namespace StudentsEducation.Web.Areas.Admin.Pages.Roles
             {
                 return NotFound();
             }
-
+            var role = await _service.GetRoleAsync(id);
+            if (role.Name == "Administrator") return RedirectToPage(Url.Content("~/Error"));
 
             await _service.DeleteRoleAsync(id);
